Cross-build to .NET 7.0 and 6.0 (#24)
Previously, scip-dotnet was only built against net7.0. Given that 6.0 is
an LTS version it makes sense to cross-build it against net6.0 as well.

There are minor differences in the snapshots between net6.0 and net7.0
(some occurrences are missing in net6.0) so I figured it was easiest to
have two separate snapshot directories. Thankfully, you can run the
tests for net6.0 and net7.0 with one `dotnet test` command. Use `dotnet
test --framework net7.0` if you only want to run the tests for a single
framework.